+++ qiskit/backends/baseprovider.py
@@ -23,7 +23,8 @@ class BaseProvider(ABC):
     def available_backends(self, *args, **kwargs):
         """
         Returns:
-            list[BaseBackend]: a list of backend available from this provider.
+            list[BaseBackend]: a list of backend instances available
+            from this provider.
         """
         pass