+++ CHANGELOG.rst
@@ -20,12 +20,18 @@ The format is based on `Keep a Changelog`_.
 
 Added
 -----
+- Retreive IBM Q jobs from server (#563, #585).
+- Add German introductory documentation (``doc/de``) (#592).
 - Add ``unregister()`` for removing previously registered providers (#584).
 - Add matplotlib-based circuit drawer (#579).
+- Adding backend filtering by least busy (#575).
 
 Changed
 -------
+- Remove backend filtering in individual providers, keep only in wrapper (#575).
 - Single source of version information (#581)
+- Bumped IBMQuantumExperience dependency to 1.9.6 (#600).
+- For backend status, `status['available']` is now `status['operational']`.
 
 Removed
 -------